+++ CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+* Added Build Report Steps view
 * Fixed detection of HDRP mixed LitShaderMode
 
 ## [0.6.1-preview] - 2021-05-11